split method, fix tests
