+++ packages/frontend/config/jest/react-markdown.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+function ReactMarkdown({ children }){
+  return <>{children}</>;
+}
+
+export default ReactMarkdown;
