+++ packages/frontend/config/jest/rehype.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+function Rehype({ children }){
+  return <>{children}</>;
+}
+
+export default Rehype;
