+++ packages/frontend/src/hooks/organisation/useOrganisation.ts
@@ -1,8 +1,7 @@
 import { ApiError, Organisation } from '@lightdash/common';
-import { useMutation, useQueries, useQuery, useQueryClient } from 'react-query';
+import { useQuery } from 'react-query';
 import { UseQueryOptions } from 'react-query/types/react/types';
 import { lightdashApi } from '../../api';
-import useToaster from '../toaster/useToaster';
 
 const getOrganisation = async () =>
     lightdashApi<Organisation>({
@@ -19,25 +18,3 @@ export const useOrganisation = (
         queryFn: getOrganisation,
         ...useQueryOptions,
     });
-
-const deleteDashboard = async (id: string) =>
-    lightdashApi<undefined>({
-        url: `/org/${id}`,
-        method: 'DELETE',
-        body: undefined,
-    });
-
-export const useDeleteOrganisationMutation = () => {
-    const { showToastError } = useToaster();
-    return useMutation<undefined, ApiError, string>(deleteDashboard, {
-        onSuccess: async () => {
-            window.location.href = '/register';
-        },
-        onError: (error) => {
-            showToastError({
-                title: `Failed to delete organisation`,
-                subtitle: error.error.message,
-            });
-        },
-    });
-};
